LayoutTests/fast/frames/message-port-postMessage-unload-event-expected.txt

Tests that MessagePort.postMessage() works from inside an iframe's unload event handler.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframe received "pagehide" event
PASS iframe received "visibilitychange" event
PASS iframe received "unload" event
PASS Parent frame received "message" event
PASS Parent frame received "message" event
PASS Parent frame received "message" event
PASS MessagePort in parent frame received message
PASS MessagePort in parent frame received message
PASS MessagePort in parent frame received message
PASS eventsReceivedByFrameWindow is 3
PASS eventsReceivedByTopWindow is 3
PASS eventsReceivedByTopWindowMessagePort is 3
PASS successfullyParsed is true

TEST COMPLETE
